添加评论计数功能，完善获取视频评论列表功能

+++ dal/model/comment.go
@@ -6,7 +6,7 @@ import (
 
 type Comment struct {
 	gorm.Model
-	VideoID int    `gorm:"index:idx_videoid;not null"`
-	UserID  int    `gorm:"index:idx_userid;not null"`
+	VideoID int64  `gorm:"index:idx_videoid;not null"`
+	UserID  int64  `gorm:"index:idx_userid;not null"`
 	Content string `gorm:"type:varchar(255);not null"`
 }
